Convert build_from into a command
